LPS-179177 Extract CommerceCurrencyExchangeRateScreenNavigationCategory methods to a CommerceCurrencyExchangeRateScreenNavigationEntry component and extends CommerceCurrencyExchangeRateScreenNavigationCategory in it
