+++ modules/apps/commerce/commerce-currency-web/src/main/java/com/liferay/commerce/currency/web/internal/frontend/taglib/servlet/taglib/CommerceCurrencyExchangeRateScreenNavigationCategory.java
@@ -14,34 +14,13 @@
 
 package com.liferay.commerce.currency.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.currency.model.CommerceCurrency;
-import com.liferay.commerce.currency.model.CommerceCurrencyConstants;
-import com.liferay.commerce.currency.service.CommerceCurrencyService;
-import com.liferay.commerce.currency.util.CommercePriceFormatter;
-import com.liferay.commerce.currency.util.ExchangeRateProviderRegistry;
 import com.liferay.commerce.currency.web.internal.constants.CommerceCurrencyScreenNavigationConstants;
-import com.liferay.commerce.currency.web.internal.display.context.CommerceCurrenciesDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
-import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
-import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -50,15 +29,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceCurrencyExchangeRateScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceCurrency> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -67,17 +42,11 @@ public class CommerceCurrencyExchangeRateScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return CommerceCurrencyScreenNavigationConstants.
-			ENTRY_KEY_COMMERCE_CURRENCY_EXCHANGE_RATE;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(
+		return language.get(
 			resourceBundle,
 			CommerceCurrencyScreenNavigationConstants.
 				ENTRY_KEY_COMMERCE_CURRENCY_EXCHANGE_RATE);
@@ -89,59 +58,7 @@ public class CommerceCurrencyExchangeRateScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_CURRENCY;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		RenderRequest renderRequest =
-			(RenderRequest)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_REQUEST);
-		RenderResponse renderResponse =
-			(RenderResponse)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_RESPONSE);
-
-		CommerceCurrenciesDisplayContext commerceCurrenciesDisplayContext =
-			new CommerceCurrenciesDisplayContext(
-				_commerceCurrencyService, _commercePriceFormatter,
-				_configurationProvider, _exchangeRateProviderRegistry,
-				_portletResourcePermission, renderRequest, renderResponse);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, commerceCurrenciesDisplayContext);
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/commerce_currency/exchange_rate.jsp");
-	}
-
 	@Reference
-	private CommerceCurrencyService _commerceCurrencyService;
-
-	@Reference
-	private CommercePriceFormatter _commercePriceFormatter;
-
-	@Reference
-	private ConfigurationProvider _configurationProvider;
-
-	@Reference
-	private ExchangeRateProviderRegistry _exchangeRateProviderRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(resource.name=" + CommerceCurrencyConstants.RESOURCE_NAME + ")"
-	)
-	private PortletResourcePermission _portletResourcePermission;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.currency.web)"
-	)
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
