+++ modules/apps/commerce/commerce-currency-web/src/main/java/com/liferay/commerce/currency/web/internal/frontend/taglib/servlet/taglib/CommerceCurrencyExchangeRateScreenNavigationEntry.java
@@ -1,0 +1,112 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.currency.web.internal.frontend.taglib.servlet.taglib;
+
+import com.liferay.commerce.currency.model.CommerceCurrency;
+import com.liferay.commerce.currency.model.CommerceCurrencyConstants;
+import com.liferay.commerce.currency.service.CommerceCurrencyService;
+import com.liferay.commerce.currency.util.CommercePriceFormatter;
+import com.liferay.commerce.currency.util.ExchangeRateProviderRegistry;
+import com.liferay.commerce.currency.web.internal.constants.CommerceCurrencyScreenNavigationConstants;
+import com.liferay.commerce.currency.web.internal.display.context.CommerceCurrenciesDisplayContext;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
+import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
+import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
+import com.liferay.portal.kernel.util.JavaConstants;
+import com.liferay.portal.kernel.util.WebKeys;
+
+import java.io.IOException;
+
+import javax.portlet.RenderRequest;
+import javax.portlet.RenderResponse;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Pedro Victor Silvestre
+ */
+@Component(
+	property = "screen.navigation.entry.order:Integer=10",
+	service = ScreenNavigationEntry.class
+)
+public class CommerceCurrencyExchangeRateScreenNavigationEntry
+	extends CommerceCurrencyExchangeRateScreenNavigationCategory
+	implements ScreenNavigationEntry<CommerceCurrency> {
+
+	@Override
+	public String getEntryKey() {
+		return CommerceCurrencyScreenNavigationConstants.
+			ENTRY_KEY_COMMERCE_CURRENCY_EXCHANGE_RATE;
+	}
+
+	@Override
+	public void render(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
+
+		RenderRequest renderRequest =
+			(RenderRequest)httpServletRequest.getAttribute(
+				JavaConstants.JAVAX_PORTLET_REQUEST);
+		RenderResponse renderResponse =
+			(RenderResponse)httpServletRequest.getAttribute(
+				JavaConstants.JAVAX_PORTLET_RESPONSE);
+
+		CommerceCurrenciesDisplayContext commerceCurrenciesDisplayContext =
+			new CommerceCurrenciesDisplayContext(
+				_commerceCurrencyService, _commercePriceFormatter,
+				_configurationProvider, _exchangeRateProviderRegistry,
+				_portletResourcePermission, renderRequest, renderResponse);
+
+		httpServletRequest.setAttribute(
+			WebKeys.PORTLET_DISPLAY_CONTEXT, commerceCurrenciesDisplayContext);
+
+		_jspRenderer.renderJSP(
+			_servletContext, httpServletRequest, httpServletResponse,
+			"/commerce_currency/exchange_rate.jsp");
+	}
+
+	@Reference
+	private CommerceCurrencyService _commerceCurrencyService;
+
+	@Reference
+	private CommercePriceFormatter _commercePriceFormatter;
+
+	@Reference
+	private ConfigurationProvider _configurationProvider;
+
+	@Reference
+	private ExchangeRateProviderRegistry _exchangeRateProviderRegistry;
+
+	@Reference
+	private JSPRenderer _jspRenderer;
+
+	@Reference(
+		target = "(resource.name=" + CommerceCurrencyConstants.RESOURCE_NAME + ")"
+	)
+	private PortletResourcePermission _portletResourcePermission;
+
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.commerce.currency.web)"
+	)
+	private ServletContext _servletContext;
+
+}
